Update Project.toml
It seems that the tests are passing fine with the latest version of SpecialFunctions, hence this suggestion. That would allow us to update to the latest version on our RHEOS library too, as we rely on both MittagLeffler and SpecialFunctions. But we could stick with older versions of SpecialFunctions too if there is a good reason to stick with version 0.10. Thanks!